moved styled tabs sub-component into balance component
