+++ src/components/wallet/balance.component.tsx
@@ -1,53 +1,10 @@
 import React, { useState } from 'react';
 
 import Grid from '@material-ui/core/Grid';
-import Tab from '@material-ui/core/Tab';
-import Tabs from '@material-ui/core/Tabs';
 import Typography from '@material-ui/core/Typography';
-import { createStyles, Theme, makeStyles, withStyles } from '@material-ui/core/styles';
+import { createStyles, Theme, makeStyles } from '@material-ui/core/styles';
 
 import { useMyriadAccount } from './wallet.context';
-
-interface StyledTabProps {
-  label: string;
-}
-
-interface StyledTabsProps {
-  value: number;
-  onChange: (event: React.ChangeEvent<{}>, newValue: number) => void;
-}
-
-const StyledTabs = withStyles({
-  flexContainer: {
-    justifyContent: 'flex-end'
-  },
-  indicator: {
-    display: 'flex',
-    justifyContent: 'center',
-    backgroundColor: '#A942E9',
-    '& > span': {
-      maxWidth: 40,
-      width: '100%'
-      // backgroundColor: '#616161'
-    }
-  }
-})((props: StyledTabsProps) => <Tabs {...props} TabIndicatorProps={{ children: <span /> }} />);
-
-const StyledTab = withStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      textTransform: 'uppercase',
-      color: theme.palette.common.white,
-      minWidth: 40,
-      fontWeight: theme.typography.fontWeightRegular,
-      fontSize: theme.typography.pxToRem(15),
-      marginRight: theme.spacing(1),
-      '&:focus': {
-        opacity: 1
-      }
-    }
-  })
-)((props: StyledTabProps) => <Tab disableRipple {...props} />);
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -82,14 +39,8 @@ const useStyles = makeStyles((theme: Theme) =>
 
 export const BalanceComponent = React.memo(function Wallet() {
   const style = useStyles();
-  const [value, setValue] = useState(0);
   const [isHidden, setIsHidden] = useState(true);
   const { state: myriadAccount } = useMyriadAccount();
-
-  const handleChange = (event: React.ChangeEvent<{}>, newValue: number) => {
-    event.preventDefault();
-    setValue(newValue);
-  };
 
   const handleIsHidden = () => {
     setIsHidden(!isHidden);
@@ -110,11 +61,6 @@ export const BalanceComponent = React.memo(function Wallet() {
           </Typography>
         </Grid>
       </Grid>
-      <StyledTabs value={value} onChange={handleChange}>
-        <StyledTab label="All" />
-        <StyledTab label="In" />
-        <StyledTab label="Out" />
-      </StyledTabs>
     </div>
   );
 });
